backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
httpx==0.27.2
redis[asyncio]==5.2.0
tenacity==9.0.0
apscheduler==3.10.4
sentence-transformers==3.3.1
pandas==2.2.3
numpy==1.26.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
pydantic-settings==2.6.1
python-dotenv==1.0.1
cryptography==43.0.3
google-genai>=1.0.0
garminconnect
cloudscraper
pytest==8.3.3
pytest-asyncio==0.24.0
